requirements.txt
argon2-cffi
cryptography
customtkinter
fastapi
gunicorn
psycopg[binary]
requests
uvicorn
